cpu_boost: Introduce interactive input boost
In the previous cpu input boost, the boost duration was difficult to weigh.
If the duration is too long, the CPU will run at high frequencies for a longer
period of time and consume more power. If the duration is too short, the boost
ends before the end of the response animation.

In the commit named "drm/atomic: Boost DDR bus when committing a new frame",
Sultan Alsawaf implemented the boost when submitting renderings to avoid
boost when no rendering request to reduce power consumption.

Therefore, we introduce interactive input boost, drawing on the practice of
Sultan Alsawaf, which only tick cpu_boost if a new frame is ready to
be rendered to the display during a fixed period of time after the input
event, and does not ramp up the frequencies if there is no rendering request.
This feature allows for a long enough boost time to cover the full response
animation without wasting too much power.